Added serilog, restored api on frontend, splited projects

+++ Models/RoomItems/Vertex.cs
@@ -4,8 +4,8 @@
     {
         public string Name { get; set; }
         public string Type { get; set; }
-        public int Y { get; set; }
-        public int X { get; set; }
+        public float Y { get; set; }
+        public float X { get; set; }
         public int Radius { get; set; }
         public string Fill { get; set; }
         public string Stroke { get; set; }
